gba: Fix boot animation skip stub
The goal is to run SWI 1 (RegisterRamReset) to clear memory and
registers, and then SWI 0 (SoftReset) to clear the last bit of BIOS
state and restart.

However, a SWI call from ARM must actually shift the syscall number by
16, because the BIOS SWI handler always looks at the second-highest byte
(the lowest byte for a THUMB SWI, the 3rd byte for an ARM SWI).

This fixes it, where previously it only called SWI 0 (never SWI 1)
because the first syscall encoded 0 in the 3rd byte.

'InitSystemStack', which sets up the stack pointer, must be called
before the syscalls can be invoked.