fpga: sim: Switch to lossless recording format
The old format was regular H.264 with AAC audio in an MP4 container.
The new one is lossless, RGB H.264, with FLAC audio, in an MKV
container.

The new format is readable by FFMPEG, VLC, mpv, etc, but not by
QuickTime (or macOS QuickLook). But it can be easily converted if
needed.