fpga: Rewrite HandheldTop SRAM controller to be synchronous
Previously, the SRAM controller didn't actually do anything to bridge
the asynchronous SRAM into the synchronous domain of the FPGA. While
this was probably fine because the specific SRAM part is quite fast
relative to the maximum system clock speed (16 MHz), it's actually quite
unsafe and up to the whims of P&R, with no particular constraints.

This commit changes the SRAM controller to better respect the actual
part's read/write cycle timings, and properly synchronize. No
constraints are added (which probably needs to happen). The latency was
increased from 0 cycles (as in, you could have a read and get the value
the same cycle) to 2 cycles (as in, the cycle you submit the read, you
wait one cycle, and then it's available ('done')) the cycle after.

Additionally, the SRAM controller is word-addressed instead of byte
addressed.

This required making a variety of changes:
* SPI interface needs address shifting to work with word-address
* SPI receiver needs to properly handle non-async memory interfaces,
  it was pretty much there (because it already does) but some of the
  semantics of SRAM were (temporarily) a bit different w.r.t. the 'done'
  signal. So maybe this didn't actually need to change.
* GBA EwramController added too much latency, would need to stall
  (fixed in another commit)
* GB needed to change to deal with synchronous multi-cycle SRAM requests