gba: Fix physical cartridge access
Works with reading ROM, and reading/writing SRAM. Doesn't yet work with
reading/writing EEPROM, and flash is untested.

This required a number of changes. Key points:
1) Putting cartridge control signals (nCS, nCS2, nRD, nWR) into
   registers, rather than generated by logic. This avoids glitching on
   the control lines in the middle of a cycle, which may have caused
   issues.
2) To do this, needed to know a cycle earlier whether a cartridge access
   was going to continue as a burst, so added a new signal to the bus (a
   little bit of a hack, but it works) to indicate whether this is going
   to happen. Ultimately this will replace the sequential signal,
   perhaps.
3) Fix some issues in the bus where the "write" signal wasn't properly
   held on the target interfaces during a multi-cycle access. So DMA,
   which does back to back reads and writes was accidentally sending the
   wrong value of the "write" signal to the cartridge in the middle.

Part of the debugging happened with a new version of EmulatedCartridge
that directly uses the nCS/nCS2/nRD/nWR control signals instead of the
'fake' faster ones. This still needs to be added later in a cleaner way.