gba: Fix cartridge DMA not being a single burst
This commit fixes the CartridgeController -- with the previous fix, DMA
accesses to the cartridge were not a single burst access. This broke
EEPROM access (and was of course, wrong).

The reason for this is that with the previous fix, a burst ends when
there's a non-sequential request, or a request to any other target.
However, DMA on the GBA maintains the burst for the entire DMA access --
this works fine when only one of the sides is the cartridge.

The way I suspect it does this is that the cartridge controller keeps a
burst active until it gets a non-sequential access *anywhere*, and the
DMA controller just keeps doing sequential accesses after the first
cartridge access (which again is kind of weird, but only the cartridge
cares about it, and you're not supposed to do cartridge to cartridge DMA
anyway).

This required changing DMA a bit, changing some of the Bus signals, and
the flow for CartridgeController.