gba: Fix EmulatedCartridge handling of ROM request aborts
With cartridge prefetch, the cartridge controller now has a way to
"abort" the current cartridge access. This just immediately sets the
nCS/nCS2 lines high and switches back to the idle state.

This works perfectly fine on a real cartridge, or on the
VerificationEmulatedCartridge (which only uses the real lines).
However, it's broken on an EmulatedCartridge, since it's using the
reqStart/end signals, and it doesn't support starting a new request
while the current one is busy (i.e. aborting).

This required some additional logic to support. It wastes time, because
we can't actually abort a memory request on the underlying controller
(SDRAM, on handheld), so we have to wait for the previous one to finish,
and then we can start another. But it's probably fine, this seems to be
a relatively rare occurrence (a couple dozen times per second, at most?)
so it is unlikely to cause any actual slowdowns.