gba: Improve prefetch timing, avoid 2 cycle penalty
This commit fixes a bug that was imposing a 2 cycle penalty on certain
accesses when prefetch was enabled. After a request to the halfword that
prefetch was currently prefetching, the controller would always go into
the next prefetch, even if a request that would abort it was pending.

This would then cause two cycles to abort the request and start it, as
opposed to starting the request right when the previous ended.

Now passing 1430/2020 mGBA timing tests