gba: Fix prefetch buffer flush and continue
There are some cases where the prefetch buffer is flushed, but prefetch
continues (specifically when requesting a halfword that's currently
being prefetched). This is a pretty common case.

In this case, the prefetch buffer needs to be flushed, because anything
left in the buffer came before the current request. However, prefetch
continues.  Previously, the register holding the address of the first
word in the prefetch buffer was incorrect (not updated after the flush),
which meant that subsequent accesses would fail to load the data in the
prefetch buffer.

This should fix the AGS Aging Test prefetch timing test.

mGBA suite now at 1510/2020

Also seems to fix crash at the beginning of a Mario Kart race