gba: Fix incorrect request address in EmuCart during abort
The "reqAddress" signal from the cartridge controller is only valid when
"reqStart" is true. So, if we get a cartridge rom request abort, we need
to register that address right then, because when we actually start the
next rom request, the address will not be valid anymore.

Fixes some issues when there are a lot of cartridge stalls, hopefully
fixes the random crashes that sometimes happen on the physical handheld.