firmware: Bypass LCD memory when writing from FPGA
The previous config was needed to work with the GBA bitstreams. After
changing the DPI driver on the FPGA to output DPI control signals
directly from registers (rather from logic), bypassing GRAM works again.

Presumably, the logic output from the DPI driver had some glitches in
it, or otherwise didn't quite meet the timing requirements. Perhaps this
was made a problem when more of the FPGA's resources were utilized with
the larger GBA logic (compared to CGB).