firmware: Move emulated/physical cartridge loading to worker thread
This commit moves the loading of physical, and especially emulated
cartridges to a background worker thread, allowing the UI thread to
remain unblocked and continue redrawing. The GBA/Gameboy bitstream
drivers have been changed to not hold the Device lock for the entire
duration of the loading, to allow the UI to respond to button presses,
etc. and remain responsive.

Redrawing still needs some work, Slint doesn't realize the entire screen
needs to be redrawn when switching bitstreams.