fpga: Re-work handling of triple framebuffering
This centralizes all handling of the framebuffer indices in a single
clock domain (the main system clock domain).

This fixes a very rare issue where, because of how the "read index" was
handled in the AV clock domain and synchronized into the system domain,
the indexes could get out of sync, leading to one of the three frames
never being written to (a flickering black frame).

I only noticed this issue once, in the GB when switching between HDMI
and internal display out.

This also adds a new "skipped frames" stat counter.