gb: Fix Serial functionality (with external clock)
The Serial functionality previously didn't actually work. The key issue
is that the shift circuitry is only supposed to be active when "enable"
in SC is true, even with external clock (slave) setting.

Link now seems to work reliably (every time), whereas before it almost
never worked, and even when it did, it sort of failed at some point.

Separately, this commit also adds a double-flip-flop synchronizer for
the two serial input signals (although that almost certainly was not the
problem).