gba: Fix issue where Reset can lock up the emulator
This was happening because the "regAccessSplit" in Bus wasn't a RegInit,
so it maintained its value over reset. If we pause and reset while
regAccessSplit=1 (easy to do if continuously making 32-bit
accesses to ROM), then regAccessSplit will stay 1 during the reset, and
the Bus will refuse to service any more requests.