fpga: Fix DirectReadCache cache fill when new request comes in
Previously, only a single port was used for the cache memory. This posed
a problem if a cache miss request finished on the same cycle as a new
request: the cache writeback would be overwritten by the cache read for
the incoming request, so the cache writeback would never happen, and the
cache wouldn't get filled, and so there would be a significant number of
cache misses.