firmware: Stop handling input events in worker thread
The worker thread has the lowest priority (lower than the UI thread). If
UI rendering is taking a long time (e.g. when navigating through a
list), the UI thread won't yield, and will starve the worker thread.

Normally this is fine and expected, but if the worker thread is also
responsible for sending the input updates, we can get stuck in a
situation where the UI is stuck in a loop (button being held down), and
we can't exit because the event that would stop it is stuck in the
worker queue.

So, pass input related events directly to UI thread, not the worker
thread. Which logically makes sense anyway.

Still some cleanup that needs to happen here (e.g. the InputManager,
while being called from the UI, is sending Button events to the UI
thread).