firmware: Use built-in ListView for FileListView
Previously, we were using an entirely custom ListView component for the
FileListView. However, large lists instantiate all of the items, which
uses a lot of memory (and can cause an OOM).

The built-in ListView has some special handling to lazily initialize
items (only those in view). However, it comes with a restriction that
you can't completely parametrize it with custom item components, you can
only have a static for-loop in the component. So this commit copies and
specializes the built in ListView for the FileListItems, and then
replicates the other stuff around it (scrollbar, etc.).