gb/gba: Fix PSG channel1 wavelength writing
PSG channel 1 is a pulse register with a sweep unit. It has the ability
to automatically adjust the frequency. When the frequency adjusts
itself, this overwrites the internal frequency (wavelength).

The CPU can also update the wavelength manually. However, a bug was
always using the frequency 'shadow register' as the wavelength of the
channel, so if you don't use sweep and just manually update the
wavelength, the sound would play wrong.

This fixes the issue on both GB and GBA.